c21112018

+++ web-client/app/shared/components/header/header.component.ts
@@ -5,6 +5,9 @@ import { TranslateService } from '@ngx-translate/core';
 import { CookieService } from 'ngx-cookie-service';
 import { CheckLoginService } from '../../../services/check-login-service';
 
+import { UserSessionService } from '../../../services/userSessionService';
+import { UserSession } from '../../../common/userSession';
+
 
 @Component({
     selector: 'app-header',
@@ -13,27 +16,25 @@ import { CheckLoginService } from '../../../services/check-login-service';
 })
 export class HeaderComponent implements OnInit {
 
-    userSession : UserSession;
+    userSession: UserSession;
     pushRightClass: string = 'push-right';
     navigationSubscription;
 
     constructor(private translate: TranslateService,
         public router: Router,
         private cookieService: CookieService,
-        @Inject(CheckLoginService) private checkLogin, 
-        @Inject(HttpCallerService) private httpCaller: HttpCallerService) {
+        @Inject(CheckLoginService) private checkLogin,
+        @Inject(HttpCallerService) private httpCaller: HttpCallerService,
+        @Inject(UserSessionService) private userSessionService: UserSessionService) {
         this.navigationSubscription = this.router.events.subscribe((val) => {
             if (val instanceof NavigationEnd && window.innerWidth <= 992 && this.isToggled()) {
                 this.toggleSidebar();
             }
         });
-        this.userSession = new UserSession();
     }
 
-    ngOnInit() { 
-
-        this.userSession.Username = this.cookieService.get("userId")
-        
+    ngOnInit() {
+        this.userSessionService.userSession.subscribe(us => this.userSession = us);
     }
 
     isToggled(): boolean {
@@ -52,15 +53,22 @@ export class HeaderComponent implements OnInit {
     }
 
     onLoggedout() {
-               
+
         this.httpCaller.callGet(
             '/logout/google',
-            () => { },
-            () => { });
-       
-        this.cookieService.delete("google_access_token")
-        this.cookieService.delete("google_refresh_token")
-        this.cookieService.delete("lastAuthTime")
+            (r) => {
+                if (r.response === 'ok') {
+                    this.cookieService.delete("google_access_token")
+                    this.cookieService.delete("google_refresh_token")
+                    this.cookieService.delete("lastAuthTime")
+                }
+                console.log('logout failed!');
+            },
+            () => {
+                console.log('logout failed!');
+             });
+
+
     }
 
     changeLang(language: string) {
@@ -78,7 +86,7 @@ export class HeaderComponent implements OnInit {
 }
 
 
-export class UserSession {
-    public Username : string;
-    public Language : string;
-}
+// export class UserSession {
+//     public Username : string;
+//     public Language : string;
+// }
